Seperated all sorting types into child class
