+++ src/main/java/com/SortContr.java
@@ -2,6 +2,10 @@ package com; /**
  * Sample Skeleton for 'template.fxml' Controller Class
  */
 
+import com.funcs.Insertion;
+import com.funcs.MergeSort;
+import com.funcs.QuickSort;
+import com.funcs.Selection;
 import javafx.animation.*;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
@@ -150,28 +154,28 @@ public class SortContr implements Initializable {
         switch (sortChoice) {
             case 1:
                 rate = 1.5 / barval;
-                sorter = new SortingFuncs(valueArray, rate);
-                sorter.insertion();
-
+                sorter = new Insertion(valueArray,rate);
+                ((Insertion)sorter).insertion();// ONLY TEMPORARY CHANGE WHEN ADDING THREAD
                 finalseq = sorter.finEffect();
+
                 break;
             case 2:
                 rate = 1.0 / barval;
-                sorter = new SortingFuncs(valueArray, rate);
-                sorter.selection();
-
+                sorter = new Selection(valueArray,rate);
+                ((Selection)sorter).selection();
                 finalseq = sorter.finEffect();
+
                 break;
             case 3:
                 rate = 3.0 / barval;
-                sorter = new SortingFuncs(valueArray, rate);
-                sorter.mergeInit();
+                sorter = new MergeSort(valueArray,rate);
+                ((MergeSort)sorter).mergeInit();
                 finalseq = sorter.finEffect();
                 break;
             case 4:
                 rate = 3.0 / barval;
-                sorter = new SortingFuncs(valueArray, rate);
-                sorter.quick();
+                sorter = new QuickSort(valueArray,rate);
+                ((QuickSort)sorter).quick();
                 finalseq = sorter.finEffect();
                 break;
             case 5:
@@ -188,6 +192,7 @@ public class SortContr implements Initializable {
 
 
     }
+
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
